+++ CHANGELOG.md
@@ -1,15 +1,17 @@
 ### ollama-c-lient-v0.0.2 - Under Dev./Testing
 
 #### date: 
-#### severity: low
+#### severity: medium
 #### bugs-fixed:
 - N/A
 #### new-features:
-- changes for alignment with [Ollama v0.9.0 changes](https://github.com/ollama/ollama/releases/tag/v0.9.0)
+- added parameter: --no-think. Alignment with [Ollama v0.9.0 changes](https://github.com/ollama/ollama/releases/tag/v0.9.0)
+- added parameters: repeat_last_n, repeat_penalty, top_k, top_p, min_p. 
 #### improvements:
 - N/A
 #### others:
 - N/A
+
 
 ### ollama-c-lient-v0.0.1